Deleting files and fixing the invoie link

+++ App-Server3/controllers/admin.js
@@ -1,4 +1,5 @@
 const mongoose = require('mongoose');
+const fileHelper = require('../util/file');
 
 const { validationResult } = require('express-validator');
 
@@ -89,7 +90,7 @@ exports.postAddProduct = (req, res, next) => {
       // const error = new Error(err);
       // error.httpStatusCode = 500;
       // return next(error);
-      console.log('line 92, admin.js ',err);
+      console.log('line 92, admin.js ', err);
     });
 };
 
@@ -156,6 +157,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       // product.imageUrl = updatedImageUrl;
@@ -192,7 +194,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if(!product){
+        return next(new Error('Product not found'))
+      } 
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
